data structure upgrade
- changed speed and attackspeed from int to float for easier balancing and finer adjustments
- lowered the player camera a bit

+++ Assets/!/Scripts/Data/CharacterData.cs
@@ -18,8 +18,8 @@ public class CharacterData
     public int health;
     public float armor;
     public int damage;
-    public int speed;
-    public int attackSpeed;
+    public float speed;
+    public float attackSpeed;
     public int cure;
 
     // weapon
@@ -59,8 +59,8 @@ public class CharacterDataDTO
     public int health;
     public float armor;
     public int damage;
-    public int speed;
-    public int attackSpeed;
+    public float speed;
+    public float attackSpeed;
     public int cure;
     public int hurt;
